stricter ints for round and valIndex (#4582)
* ints: change some ints to int32

- this is to help reduce the overall size of some proto prs.


* add safemath for int32

* add safemath for subtraction

* add go doc to safemath funcs

* move panics to safe math, add checkconvert func

* make safe convert accept int64

+++ consensus/common_test.go
@@ -67,16 +67,16 @@ func ResetConfig(name string) *cfg.Config {
 // validator stub (a kvstore consensus peer we control)
 
 type validatorStub struct {
-	Index  int // Validator index. NOTE: we don't assume validator set changes.
+	Index  int32 // Validator index. NOTE: we don't assume validator set changes.
 	Height int64
-	Round  int
+	Round  int32
 	types.PrivValidator
 	VotingPower int64
 }
 
 var testMinPower int64 = 10
 
-func newValidatorStub(privValidator types.PrivValidator, valIndex int) *validatorStub {
+func newValidatorStub(privValidator types.PrivValidator, valIndex int32) *validatorStub {
 	return &validatorStub{
 		Index:         valIndex,
 		PrivValidator: privValidator,
@@ -166,15 +166,15 @@ func (vss ValidatorStubsByPower) Less(i, j int) bool {
 func (vss ValidatorStubsByPower) Swap(i, j int) {
 	it := vss[i]
 	vss[i] = vss[j]
-	vss[i].Index = i
+	vss[i].Index = int32(i)
 	vss[j] = it
-	vss[j].Index = j
+	vss[j].Index = int32(j)
 }
 
 //-------------------------------------------------------------------------------
 // Functions for transitioning the consensus state
 
-func startTestRound(cs *State, height int64, round int) {
+func startTestRound(cs *State, height int64, round int32) {
 	cs.enterNewRound(height, round)
 	cs.startRoutines(0)
 }
@@ -184,7 +184,7 @@ func decideProposal(
 	cs1 *State,
 	vs *validatorStub,
 	height int64,
-	round int,
+	round int32,
 ) (proposal *types.Proposal, block *types.Block) {
 	cs1.mtx.Lock()
 	block, blockParts := cs1.createProposalBlock()
@@ -221,7 +221,7 @@ func signAddVotes(
 	addVotes(to, votes...)
 }
 
-func validatePrevote(t *testing.T, cs *State, round int, privVal *validatorStub, blockHash []byte) {
+func validatePrevote(t *testing.T, cs *State, round int32, privVal *validatorStub, blockHash []byte) {
 	prevotes := cs.Votes.Prevotes(round)
 	pubKey, err := privVal.GetPubKey()
 	require.NoError(t, err)
@@ -259,7 +259,7 @@ func validatePrecommit(
 	t *testing.T,
 	cs *State,
 	thisRound,
-	lockRound int,
+	lockRound int32,
 	privVal *validatorStub,
 	votedBlockHash,
 	lockedBlockHash []byte,
@@ -307,7 +307,7 @@ func validatePrevoteAndPrecommit(
 	t *testing.T,
 	cs *State,
 	thisRound,
-	lockRound int,
+	lockRound int32,
 	privVal *validatorStub,
 	votedBlockHash,
 	lockedBlockHash []byte,
@@ -413,7 +413,7 @@ func randState(nValidators int) (*State, []*validatorStub) {
 	cs := newState(state, privVals[0], counter.NewApplication(true))
 
 	for i := 0; i < nValidators; i++ {
-		vss[i] = newValidatorStub(privVals[i], i)
+		vss[i] = newValidatorStub(privVals[i], int32(i))
 	}
 	// since cs1 starts at 1
 	incrementHeight(vss[1:]...)
@@ -462,7 +462,7 @@ func ensureNoNewTimeout(stepCh <-chan tmpubsub.Message, timeout int64) {
 		"We should be stuck waiting, not receiving NewTimeout event")
 }
 
-func ensureNewEvent(ch <-chan tmpubsub.Message, height int64, round int, timeout time.Duration, errorMessage string) {
+func ensureNewEvent(ch <-chan tmpubsub.Message, height int64, round int32, timeout time.Duration, errorMessage string) {
 	select {
 	case <-time.After(timeout):
 		panic(errorMessage)
@@ -482,7 +482,7 @@ func ensureNewEvent(ch <-chan tmpubsub.Message, height int64, round int, timeout
 	}
 }
 
-func ensureNewRound(roundCh <-chan tmpubsub.Message, height int64, round int) {
+func ensureNewRound(roundCh <-chan tmpubsub.Message, height int64, round int32) {
 	select {
 	case <-time.After(ensureTimeout):
 		panic("Timeout expired while waiting for NewRound event")
@@ -501,13 +501,13 @@ func ensureNewRound(roundCh <-chan tmpubsub.Message, height int64, round int) {
 	}
 }
 
-func ensureNewTimeout(timeoutCh <-chan tmpubsub.Message, height int64, round int, timeout int64) {
+func ensureNewTimeout(timeoutCh <-chan tmpubsub.Message, height int64, round int32, timeout int64) {
 	timeoutDuration := time.Duration(timeout*10) * time.Nanosecond
 	ensureNewEvent(timeoutCh, height, round, timeoutDuration,
 		"Timeout expired while waiting for NewTimeout event")
 }
 
-func ensureNewProposal(proposalCh <-chan tmpubsub.Message, height int64, round int) {
+func ensureNewProposal(proposalCh <-chan tmpubsub.Message, height int64, round int32) {
 	select {
 	case <-time.After(ensureTimeout):
 		panic("Timeout expired while waiting for NewProposal event")
@@ -526,7 +526,7 @@ func ensureNewProposal(proposalCh <-chan tmpubsub.Message, height int64, round i
 	}
 }
 
-func ensureNewValidBlock(validBlockCh <-chan tmpubsub.Message, height int64, round int) {
+func ensureNewValidBlock(validBlockCh <-chan tmpubsub.Message, height int64, round int32) {
 	ensureNewEvent(validBlockCh, height, round, ensureTimeout,
 		"Timeout expired while waiting for NewValidBlock event")
 }
@@ -566,12 +566,12 @@ func ensureNewBlockHeader(blockCh <-chan tmpubsub.Message, height int64, blockHa
 	}
 }
 
-func ensureNewUnlock(unlockCh <-chan tmpubsub.Message, height int64, round int) {
+func ensureNewUnlock(unlockCh <-chan tmpubsub.Message, height int64, round int32) {
 	ensureNewEvent(unlockCh, height, round, ensureTimeout,
 		"Timeout expired while waiting for NewUnlock event")
 }
 
-func ensureProposal(proposalCh <-chan tmpubsub.Message, height int64, round int, propID types.BlockID) {
+func ensureProposal(proposalCh <-chan tmpubsub.Message, height int64, round int32, propID types.BlockID) {
 	select {
 	case <-time.After(ensureTimeout):
 		panic("Timeout expired while waiting for NewProposal event")
@@ -593,15 +593,15 @@ func ensureProposal(proposalCh <-chan tmpubsub.Message, height int64, round int,
 	}
 }
 
-func ensurePrecommit(voteCh <-chan tmpubsub.Message, height int64, round int) {
+func ensurePrecommit(voteCh <-chan tmpubsub.Message, height int64, round int32) {
 	ensureVote(voteCh, height, round, types.PrecommitType)
 }
 
-func ensurePrevote(voteCh <-chan tmpubsub.Message, height int64, round int) {
+func ensurePrevote(voteCh <-chan tmpubsub.Message, height int64, round int32) {
 	ensureVote(voteCh, height, round, types.PrevoteType)
 }
 
-func ensureVote(voteCh <-chan tmpubsub.Message, height int64, round int,
+func ensureVote(voteCh <-chan tmpubsub.Message, height int64, round int32,
 	voteType types.SignedMsgType) {
 	select {
 	case <-time.After(ensureTimeout):
